Refactored 'main.jsx' component
