+++ src/components/main/main.jsx
@@ -2,8 +2,6 @@ import React from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
 
-import {ActionCreator} from "../../store/action";
-
 import {movieType} from "../../types";
 
 import withMoviesList from "../../hocs/with-movies-list/with-movies-list";
@@ -15,15 +13,7 @@ import ShowMore from "../show-more/show-more";
 const MoviesListWrapped = withMoviesList(MoviesList);
 
 const Main = (props) => {
-  const {
-    movie,
-    movies,
-    moviesByGenre,
-    shownMovies,
-    activeGenre,
-    onGenreClick,
-    onShowMoreClick
-  } = props;
+  const {movie, movies, moviesByGenre, shownMovies} = props;
 
   return (
     <React.Fragment>
@@ -88,8 +78,6 @@ const Main = (props) => {
 
           <GenresList
             movies={movies}
-            activeGenre={activeGenre}
-            onGenreClick={onGenreClick}
           />
 
           <MoviesListWrapped
@@ -100,7 +88,6 @@ const Main = (props) => {
           {shownMovies < moviesByGenre.length &&
             <ShowMore
               shownMovies={shownMovies}
-              onShowMoreClick={onShowMoreClick}
             />
           }
         </section>
@@ -125,30 +112,17 @@ const Main = (props) => {
 
 const mapStateToProps = (state) => {
   return {
-    activeGenre: state.activeGenre,
     moviesByGenre: state.moviesByGenre,
     shownMovies: state.shownMovies,
   };
 };
-
-const mapDispatchToProps = (dispatch) => ({
-  onGenreClick(activeGenre) {
-    dispatch(ActionCreator.changeMoviesGenre(activeGenre));
-  },
-  onShowMoreClick(shownMovies) {
-    dispatch(ActionCreator.showMoreMovies(shownMovies));
-  }
-});
 
 Main.propTypes = {
   movie: movieType,
   movies: PropTypes.arrayOf(movieType),
   moviesByGenre: PropTypes.arrayOf(movieType),
   shownMovies: PropTypes.number.isRequired,
-  activeGenre: PropTypes.string.isRequired,
-  onGenreClick: PropTypes.func.isRequired,
-  onShowMoreClick: PropTypes.func.isRequired,
 };
 
 export {Main};
-export default connect(mapStateToProps, mapDispatchToProps)(Main);
+export default connect(mapStateToProps, null)(Main);
